Create binary release workflow
This adds a GHA workflow to release dbc binaries for multiple platforms when a Golang-style `vX.Y.Z` type tag is pushed. The binaries are cross-compiled on CI, tar-gz'd up with the README.md and LICENSE.md, and a new GitHub Release is created for the tag with the archives are attached.

See https://github.com/amoeba/dbc/releases/tag/v0.1.0 for an example of running this on my fork.

Each archive has this structure,

```sh
$ tree dbc-darwin-arm64-v0.1/
dbc-darwin-arm64-v0.1/
├── dbc
├── LICENSE
└── README.md
```

A big thing missing here is included detailed/useful release notes would be a manual step so that can be a follow-up.

Closes https://github.com/columnar-tech/dbc/issues/29.