fix(cmd/dbc): catch nil msg.Version and add test (#59)
This fixes an error where an existing manifest on disk is missing the version field and we to try to install a new driver with the same name. e.g., invalid bigquery.toml and we run dbc install bigquery. This also extends the current test table to support a setup routine and a check on the output.