Improve dbc.toml not found errors
This is a thing I'm starting to notice more and more with Go error handling: that you can wrap errors in errors and end up with verbose and impenetrable errors.

In this case, we only want to tell the user dbc.toml doesn't exist and only advise them to run dbc init if that's the case. All other cases, we want to show the potentially more serious error