fix: add test to cover behavior for invalid manifests (#48)
Fixes an edge case where we get confused about what fields to put in a
manifest. When we package a driver, we include an all-caps `MANIFEST`
file which has a `Files` table that points to the shared library living
in the same package. When dbc installs a package, it converts what's in
`Files` into what goes in `Driver.shared`. If we goof up and put a
`Driver.shared` table in `MANIFEST`, `dbc uninstall` was mistakenly
deleting _all_ drivers in the config level.

This patch catches this case and handles it by skipping. It's good to be defensive
here because we'll be asking people to crate `MANIFEST` files who aren't
us so mistakes will happnen.

Closes https://github.com/columnar-tech/dbc/issues/29