docs: first draft of documentation site (#33)
This is a very first-draft draft of what I'm thinking of for the dbc
docs site. I've caught a number of bugs which I've created issues for in
the [0.0.1 milestone](https://github.com/columnar-tech/dbc/milestone/1).
I expect there will be a lot of changes once I put fresh eyes on it and
others take a look.

---------

Co-authored-by: David Li <li.davidm96@gmail.com>
Co-authored-by: Ian Cook <ianmcook@gmail.com>
Co-authored-by: Matt Topol <zotthewizard@gmail.com>