fix: add workaround for driver manager venv path bug (#102)
This works around a bug we identified with the 1.8.0 driver managers
when using virtual environments. They are supposed to locate drivers in
`$VIRTUAL_ENV/etc/adbc/drivers` but they look in
`$VIRTUAL_ENV/etc/adbc/`. There may be a patch release soon to fix this
but, in the mean time, we can work around this by symlinking manifests
from `$VIRTUAL_ENV/etc/adbc/drivers` to `$VIRTUAL_ENV/etc/adbc/`. The
manifests use absolute paths so it's fine.

This PR also adds an integration workflow which is how I tested my
change. I don't think we want to run it on every single commit to the
repo so we might discuss here when we might want to run something like
this or if we want to run it in another repo.