feat(docs): add subcommand for opening documentation (#186)
Adds the implementation and documentation for a new subcommand, `docs`.

If you run `dbc docs` with no driver, the main dbc docs site opens in
your default browser. If you run `dbc docs DRIVER`, we look for that
driver in the driver registry (the CDN). If not found, we error. If
found, we look in a list of temporary, fallback URLs for a docs URL and
open that. For drivers that exist in the registry but have no fallback
URL, we error.

Closes https://github.com/columnar-tech/dbc/issues/176