fix(search,windows): show installed location for all config levels (#207)
Fixes an issue where the driver install location wasn't showing up when
a user runs `dbc search -v` on Windows. This was only impacting user and
system level installs and the reason we didn't catch this is that we
don't test those levels in CI.

This PR updates the Windows code paths that set the FileInfo member on
the DriverInfo struct so it's set to the registry root for the
appropriate level. We just weren't setting it before for registry user
and system levels. I went with the registry key root instead of the
location of the DLL because the driver is really installed to the
registry and we don't want users thinking they can uninstall a driver
from %APPDATA% or something.

Here's an example of what this looks like on my system:

```
• sqlite
   Title: ASF SQLite Driver
   Description: An ADBC driver for SQLite developed under the Apache Software Foundation
   License: Apache-2.0
   Installed Versions:
    ╰── 1.8.0
        ├── system => HKEY_LOCAL_MACHINE\SOFTWARE\ADBC\Drivers
        ├── env => C:\Users\Bryce\src\columnar-tech\dbc
        ╰── user => HKEY_CURRENT_USER\SOFTWARE\ADBC\Drivers
   Available Versions:
```

Closes https://github.com/columnar-tech/dbc/issues/203