feat(registry): multi-registry support (#225)
Enable `GetDriversList` to use multiple registries as source for the
driver list, defaulting to both the public and dbc-cdn-private
registries. Will ignore a registry if it receives a 401 Unauthorized
response from the index.yaml request.

This does not yet provide a way to configure the available registries,
I'll add that in a follow-up