docs: fix logo rendering issue in social preview images (#237)
We use the [social
plugin](https://squidfunk.github.io/mkdocs-material/plugins/social/) to
generate social preview images. By default, the logo it shows in the top
right of the cards is just whatever you set in the `theme.logo` field.
The SVG logo doesn't render correctly and gets clipped.

This patch overrides the default behavior to use the 512x512 logo we
generated for site meta.

Closes #209